Re-implement more tests, delete old test file
