+++ contracts/governance/extensions/GovernorCountingFractional.sol
@@ -36,7 +36,7 @@ abstract contract GovernorCountingFractional is Governor {
      */
     // solhint-disable-next-line func-name-mixedcase
     function COUNTING_MODE() public pure virtual override returns (string memory) {
-        return "support=bravo&quorum=for,abstain";
+        return "support=bravo&quorum=for";
     }
 
     /**
